DAOS-15262 vos: Fix probe issue in vos iterator (#13918)
The checksum scrubber uses vos_iterate to scan everything in the tree, whether
visible or not whereas other service ULTs take great care to avoid any
epoch range conflict with aggregation and discard services.  This means
that any yield in the iterator needs to check to see if the parent
objects or keys are still present before continuing.

This function adds a check before probing in the current level for this
condition and backs out to higher level to continue, if necessary.  This
check is not necessary for visible iterators or aggregation iterators
because there are external guarantees that parent trees will never be
removed mid-iteration.

Signed-off-by: Jeff Olivier <jeffolivier@google.com>