DAOS-15417 build: Fix intermittent issue with libfuse build.
Meson "setup" sets up a package for buidlding, meson configure
sets a configuration option, but does not do the setup.
Previously our code would do setup, then configure which would
set configuration options but not apply them.

Ninja has file age checking built-in so if the config file was
older than the build file then it would re-run setup to apply
the correct config, and this was happening most times so
the build would work, but occasionally the file timestamps would
be the same so the check would not fail and the build would be run
without the configuration options applied.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>