DAOS-15499 dtx: re-execute condition punch
For condition punch, the target to be punched maybe only exists on
some shards. Under such case, although the condition punch will be
successful, there will be related DTX entry for the shards without
the target to be punched. Then two potential trouble for that:

1. If DTX resync happens before such DTX committed, then such DTX
   may be regarded as non-prepared one and aborted by wrong.

2. If above non-existence case happens on the DTX leader, the flag
   committable on the fake entry may misguide subsequent operation.

So for condition punch, the DTX leader needs to know whether there
are some DTX participants failed to execute the condition check or
not. If yes, then re-execute the punch unconditionally after it is
verified by the others that related target (to be punched) exists.

The patch also does some code cleanup.

Signed-off-by: Fan Yong <fan.yong@intel.com>