DAOS-15441 test: increase ior_timeout for with_ior
The original 120S possibly not enough, in a failed case it is observed
that an RPC to killed rank 3 got 60 seconds timeout that is expected,
left 60 Second possibly not enough to run the whole IOR as need to wait
timeout and refresh pool map and retry.
Increase whole test's timeout accordingly.

Test-tag: test_rebuild_with_ior
Test-repeat: 5

Required-githooks: true

Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>