DAOS-15575 dfs: replace DAOS_TX_NONE with dfs th
This patch lays the ground work for adding support for snapshot
to dfs in subsequent patches. It does not change anything from
a functionality standpoint and just replaces all occurences of
DAOS_TX_NONE in libdfs with dfs->th which can only be set to
DAOS_TX_NONE for now.

A dfs::th_epoch field is added to the global dfs handle to record
the epoch associated with the handle (can only be set to
DAOS_EPOCH_MAX for now, which means DAOS_TX_NONE).

Change dfs_dir_anchor_set() API to pass the dfs mount.

Signed-off-by: Johann Lombardi <johann.lombardi@gmail.com>