DAOS-15256 object: coll ops against object with sparse ranks (#13881)
We define the degree of rank# sparse for an object as:
(the count of ranks on which the object shards reside) /
(the count of the ranks in the pool map)

For example: the pool is built on the 10-ranks set:
[1, 3, 5, 7, 9, 10, 11, 12, 13, 14]

Case1: If the object shards reside on the ranks set [1, 3, 9, 5],
then its sparse degree is equal to 4 / 10 = 0.4

Case2: If the object shards reside on the ranks set [12, 10, 13],
then its sparse degree is equal to 3 / 10 = 0.3

Based on such sparse degree, when handle collective operation
on client, we will prepare related target bitmap as following:

If the sparse degree is not less than 0.35, then pre-allocate continuous
array (that indexed with the rank#) to record related targets bitmap for
the collective operation. It is the most efficient solution for locating
the rank slot with the rank#. It may waste some DRAM space, but that is
temporary before sending out the collective RPC. The percentage of waste
DRAM will not be more than 65%.

Otherwise, if the sparse degree is less than 0.35, then use btree as the
basic data structure to record related targets bitmap. It will introduce
some overhead for data comparing and movement when locate the rank slot
with the rank#, but it avoids occupying too much DRAM space, especially
when the sparse degree is very small.

In practice, before completely scanning the object layout, we cannot
exactly know on which ranks the object shards reside. So we need to
calculate the sparse degree based on some estimation. For example,
when generate the object layout, record the maximum and the minimum
rank# on which related object shards reside, then use the width of
the ranks range to estimate the ranks count for the given object.
Because these ranks may be not continuous, then the count may be
over-estimated. Consider that, we can adjust the boundary of sparse
degree for above ranks array and btree solutions, such as 0.35 => 0.45,
or similarly.

Signed-off-by: Fan Yong <fan.yong@intel.com>