DAOS-15514 container: fix container destroy failure
The container might be opened when the target is up,
but changed to down when closing. We need to attempt to
close down/downout targets regardless; it won't take any action
if it was not opened before. Failure to properly close it will
result in container destruction failing with EBUSY. (See DAOS-15514)

Test-tag: test_osa_online_reintegration
Required-githooks: true
Signed-off-by: Wang Shilong <shilong.wang@intel.com>