DAOS-623 build: Add --build-deps=fetch option
Allows one to split the download from the build. It is
expected to be used with --build-deps=build-only which
skips the download phase.

Change build-only to be consistent with only option in
that it exits after prerequisite build.

Fix an issue with ALT_PREFIX, resolve whole path
Set pkg-config path when checking prebuilt_path

Required-githooks: true

Change-Id: Icf0023b66ea8f4bd02fe054b8509bc10a1c6c957
Signed-off-by: Jeff Olivier <jeffolivier@google.com>