DAOS-15499 dtx: persistently store DTX entry for noop
It is possible that some transaction modifies something on some DTX
participants but nothing on the other participants. Under such case,
for the DTX participants on which nothing is modified, we also need
to store the DTX persistently. Otherwise, the subsequent DTX resync
may not find it as to regard it as failed transaction and abort it.

Signed-off-by: Fan Yong <fan.yong@intel.com>