DAOS-15614 tests: more wait when exclude rank during CR test
Currently, we do not have some efficient way to make the test
scripts to be aware of whether rank is really excluded during
CR test. So have to wait some relative long time (30 seconds).
Since there are very few use cases for that, it will not much
increase the whole test time.

Signed-off-by: Fan Yong <fan.yong@intel.com>