DAOS-XXX engine: Add single engine multi-socket mode
For simplicity, target and helper count must be a multiple
of the number of sockets and there must be fewer helpers
than targets.

Let's say K is the number of xstreams of each type that map to
a numa node. They are assigned in order with K going to the
first node, K to the second, etc.

Enabled with DAOS_MULTISOCKET=1

If not enabled, it should work as it did before.

Required-githooks: true

Change-Id: Ia7801289f4cc9e2cca4649d5132a1a5c0ea2f299
Signed-off-by: Jeff Olivier <jeffolivier@google.com>