DAOS-12648: control: Provide workaround for SPDK issue 2683
In SPDK issue 2683, SPDK incorrectly determines that an AMD iommu
does not support virtual addressing.  As a result, DAOS engines
cannot run as non-root on an AMD server.  Change DAOS to specify
iova_mode = "va" when calling spdk_env_init() when running as
non-root.  This is a workaround until SPDK issue 2683 is fixed.

Signed-off-by: Kevan Rehm <kevan.rehm@hpe.com>