DAOS-15514 container: fix container destroy failure (#14108)
The container might be opened when the target is up,
but changed to down when closing. We need to attempt to
close down/downout targets regardless; it won't take any action
if it was not opened before. Failure to properly close it will
result in container destruction failing with EBUSY. (See DAOS-15514)

Required-githooks: true

Change-Id: I82e7c507d5f165057d562455e31d7b02d19e1815
Signed-off-by: Wang Shilong <shilong.wang@intel.com>