DAOS-15672 rebuild: Fix pool destroy hangs (#14183)
See the Jira ticket for the long story. When releasing resources, for
instance, for stopping a pool, we should attempt on all targets, because
the latest version of the pool map may differ from the version used to
acquire the resources. (And in some cases, we no longer have the
original pool map version in the context.)

Signed-off-by: Li Wei <wei.g.li@intel.com>