DAOS-15329 cq: Disable debug locking macros for coverity.
The coverity tool gets confused about the use of assert in the
debug version of the logging macros and can think a lock is
being unlocked twice which it reports as a API usage error.

Disable the complex macros for coverity to reduce the instances
of false positives in the tool.

Fixes coverity ID 1975167 and others.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>