DAOS-623 dfuse: Update dfuse thread names.
When setting these previously I thought they only appeared in
debugger output so they have names which are only meaningful
in that context, but the thread names are also visable in
ps and top and having a process called "main" does not make
sense here.

Do not rename the main dfuse thread, and use a dfuse prefix
for other thread names.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>