DAOS-15549 control: Pause group update if provider is changed (#14182)
If the system fabric provider is updated, it is possible that group
updates could be sent with stale information (a mix of old and new
URIs). Engines may attempt to begin communicating with peers whose URIs
have not been updated to the new provider. We need to prevent this.

This patch temporarily pauses group updates when it detects a fabric
provider change in order to allow all known ranks to rejoin. Afterward
it triggers the group update to the engines.

Signed-off-by: Kris Jacque <kris.jacque@intel.com>