DAOS-15076 engine: Restricting Heap Usage to Sorted Resent RPCs (#14268)
In certain rare cases, RPC throttling may occur.
The initial motivation behind this change is to utilize the heap
exclusively for sorted resent RPCs and the FIFO list for regular
fetch and update requests. This strategic allocation aims to avoid
potential performance impacts that could result from maintaining a heap
in the critical hot path.

Signed-off-by: Wang Shilong <shilong.wang@intel.com>