DAOS-15691 test: CR orphan container - Don't automatically fail when … (#14254)
In an MD-on-SSD cluster, /mnt/daos?/<pool_uuid>/vos-0 isn't found after
stopping a server because it's unmounted by control plane. This
means ddb can't be used, which means we can't verify that the container
still exists in shard after running the fault injection, or the container
was removed from shard by the checker. Update the logic so that if vos-0
isn't found, skip these checks.

Signed-off-by: Makito Kano <makito.kano@intel.com>