DAOS-15824 vos: per-pool based pre-allocated DTX LRU array
The DTX LRU array is pre-allocated in DRAM for active DTX entries. Currently,
it is per-container based. In theory, for each opened container, the count of
such pre-allocated DTX entries can be up to 1^20 at most. Each DTX entry will
occupy 264 bytes. Then the DRAM usage will be 264MB per-container shard. From
server perspective, under the worst case, the total DRAM usage will be:

264MB * targets_per_engine * engines_per_server

For large system with a lot of pools and containers, that will be terrible.

This patch changes the DTX LRU array from container to pool. If there are a
lot of containers (assume the count as N) within the same pool, then it can
reduce the server DRAM consumption to 1 / N for the best case, that is much
helpful.

Signed-off-by: Fan Yong <fan.yong@intel.com>