DAOS-15589 test: Remove usage of hard-coded mount point in CR tests
In pool_list_consolidation.py and pool_membership.py, mount point
is hard-coded as /mnt/daos0. This causes an issue in MD-on-SSD
clusters because the mount point there is /mnt/daos. Update the
test to obtain the current mount point using the test harness.

Skip-unit-tests: true
Skip-fault-injection-test: true
Skip-func-hw-test-medium-md-on-ssd: false
Test-tag: PoolListConsolidationTest PoolMembershipTest

Signed-off-by: Makito Kano <makito.kano@intel.com>