DAOS-15685 control: Make failure clear when mountpt goes missing (#14293)
In the case that a normally persistent mountpoint directory goes
missing, e.g. after a catastrophic failure, the control log is not
clear about the reason for the mount failure. The message indicating
that the superblock is missing is particularly confusing and could
result in unnecessary data loss if the admin decides they need to wipe
PMem as a result.

Fix by stopping engine start-up if PMem device has ext4 filesystem
signature and the target mount-point location doesn't exist. The
engine process will not be started in this situation and a clear fault
description and resolution message will be displayed. No "waiting for
metadata format" log entry will be printed.

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>