DAOS-15589 test: If MD-on-SSD cluster, pass the test
If MD-on-SSD cluster is used, if servers are stopped, mount point is
removed by the control plane. This means we can't test the feature of
the checker that fixes the issue introduced by removing the files in
the mount point. Thus, pass the test in such case.

Refactor the code.

Skip-unit-tests: true
Skip-fault-injection-test: true
Skip-func-hw-test-medium-md-on-ssd: false
Test-tag: PoolListConsolidationTest PoolMembershipTest
Signed-off-by: Makito Kano <makito.kano@intel.com>