DAOS-15615 test: Clear existing tmpfs mount points before running tests (#14295)
Some test occasionally fail to start servers due to insufficient
available memory in CI due to left over DAOS mount points from a
previous test.  Adding an option to launch.py to provided a filter,
which if specified, will be used to umount and remove the directory for
any mounted tmpfs filesystems matching the filter. When using --mode=ci
the filter will be set to /mnt/daos.

Signed-off-by: Phil Henderson <phillip.henderson@intel.com>