DAOS-15850 control: Pass up errors when failing to read cert dir
Previously we returned a Fault indicating bad permissions whenever we
failed to read the client cert directory. This could be misleading if,
for example, the directory was a file or didn't exist at all.

Also refactored some older unit tests to not depend on in-source
directories with assumed permissions.

Features: control security

Required-githooks: true

Signed-off-by: Kris Jacque <kris.jacque@intel.com>