DAOS-16010 control: add a cloud target for support logs
Existing implementation rsync's the support logs between hosts in the cluster.
This commit adds a "-C" option that causes the servers to write copy
their results to a cloud bucket and directory.
"-C gcs://bucket/dir1/dir2"

There is also a fix that allow collect-log to work in "insecure mode"

Signed-off-by: Chris Davis <chrd@google.com>