DAOS-15589 test: Pass the CR tests if contents of mount point are removed on MD-on-SSD cluster (#14350) (#14525)
In pool_list_consolidation.py and pool_membership.py, the tests verify the feature of the checker after introducing the faults into the files in the mount point. However, in MD-on-SSD cluster, the contents of the mount point are removed by the control plane when the server is stopped. This means we can't test the checker, so pass the test.

Also obtain the mount point using server_manager and refactor.

Signed-off-by: Makito Kano <makito.kano@intel.com>