DAOS-15956 control: Limit race between taking leadership and join (#14541)
In very limited situations, it is possible to encounter a race condition
between an MS replica stepping up as leader and engines beginning to
join the system. In that case, the system fabric provider property may
not yet be set, causing the engine join to fail.

This change shrinks the gap where the race can occur. It also returns
a retryable error on the small chance that this scenario is hit.

Signed-off-by: Kris Jacque <kris.jacque@intel.com>