DAOS-16010 control: add support to override rync binary and add args
If the config "support->file_transfer_exec" is set. This program/binary
will be used instead of rsync in a  "dmg support collect-log"
This allows an example, to execute an enternal tool like gcloud to copy
results to a GCS bucket.
Also a fix for allowing the collect command to work in insecure mode is
included here.

Change-Id: I3d3feadf4383fabfd981d12b10c7aad25d9ac850
Signed-off-by: Chris Davis <chrd@google.com>