DAOS-15998 pool: Let all pool_svc rest on ds_pool
Currently, only leader pool_svc objects depend on ds_pool objects. This
unintentionally allows the creation of a PS replica even when the local
ds_pool object is being stopped, leading to the hang in the Jira ticket.
This patch lets every follower pool_svc object depend on the ds_pool
object too, so that if the latter is stopping or doesn't exist, the
pool_svc object won't be created successfully.

  - Initialize pool_svc.ps_pool when allocating pool_svc, rather than
    when stepping up.
      - Move ps_pool up in pool_svc, as it's no longer a leader-only
	field.
      - Modify init_svc_pool accordingly and rename it to
	update_svc_pool.

  - Move ds_pool_svc_stop into ds_pool_stop, because we want to set
    ds_pool.sp_stopping before stopping the PS (if any).
      - Move the PS start code into ds_pool_start for symmetry.
      - Clean up the error handling in and for ds_pool_svc_stop.

  - Change ds_pool_stop_all to stop all pools concurrently.
      - Remove ds_rsvc_stop_all, as it's no longer used anywhere.

Features: pool
Signed-off-by: Li Wei <wei.g.li@intel.com>
Required-githooks: true