DAOS-9247 control,bio: Add PCIe link speed and width to NVMe health stats (#14395)
Add NVMe PCIe link speed and width details to the NVMe health stats
returned from dmg storage query list-devices --health and dmg storage
scan --nvme-health commands. The PCIe config space is fetched within
the engine SPDK process (which is functional even when NVMe device is
behind a VMD and bound to a userspace driver) and passed back to the
control-plane over dRPC as a byte-string. The byte-string is formatted
and passed to lspci to convert to human readable text. The output text
is parsed and relevant entries converted into health stat fields to be
printed in dmg output.

Required-githooks: true

Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>