b/352550886 control: Allow Jumpbox to contact servers regardless of version
The upstream compatibility policy requires that admin clients
(e.g. dmg, Jumpbox) must be the same version as daos_server. In
our environment, this hard requirement may not make as much sense.

This commit adds a small workaround that allows the control client
to masquerade as whichever version the server reports.

Required-githooks: true

Change-Id: I2a170bddf3ee62ede082da10939510880a6702ac
Signed-off-by: Michael MacDonald <mjmac@google.com>