DAOS-16234 test: add a 1 ms sleep in the dfs test
In some cases one of the servers has a time stamp that is some nanoseconds different from the client. If that happens we can see this issue in the mtime and ctime test to make it look like we went back in time. Because the ctime reported from the dfs_osetattr() is the one on the client node, and it’s comparing to an mtime value that was obtained from an hlc obtained from the server side. so I will update the test to sleep for 10 ms to avoid such issue.

Quick-Functional: true
Test-repeat: 10
Test-tag:test_daos_dfs_unit
Skip-func-hw-test-large-md-on-ssd: false
Priority: 2

Required-githooks: true

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>