DAOS-16103 control: Avoid deadlock during leadership change
This patch addresses a corner case where a leadership change may
occur while a pool is in the Destroying state. During checkPools,
if the pool is in a Destroying state, the MS will attempt to
destroy the pool. The top level PoolDestroy method attempted to
wait for leadership step-up to finish, but was needed in order for
step-up to complete.

It is okay to skip the leadership check for PoolDestroy in this
case. Leadership was already checked at the beginning of checkPools.

Test-tag: test_ms_failover
Test-repeat: 10

Required-githooks: true

Signed-off-by: Kris Jacque <kris.jacque@intel.com>