DAOS-16234 test: add a 10 ms sleep in the dfs time test
In some cases one of the servers has a time stamp that is some microseconds different from the client. If that happens we can see this issue in the mtime and ctime test to make it look like we went back in time. Because the ctime reported from the dfs_osetattr() is the one on the client node, and it’s comparing to an mtime value that was obtained from an hlc obtained from the server side. So I will update the test to sleep for 10 ms to avoid such issue.

Test-tag:test_daos_dfs_unit

Required-githooks: true

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>