DAOS-16151 cart: add no-sync option to self_test (#14697)
- Added --no-sync option (-n) to self_test. When set will skip
pre-ping of all ranks in the group before running the actual self_test

- Usages of self_test with daos_agent env (-u option) changed to automatically disable
sync.

Rationale: crtu_wait_for_ranks() which performs sync was added for scenarios when
self_test is started in a stand-alone mode, which then allows self_test to block until all test
servers are up. In the production case (using daos agent), this step is not needed and can lead
to self_test aborting on a dead target that is not related to the one attempting to be tested.

Required-githooks: true

Signed-off-by: Alexander A Oganezov <alexander.a.oganezov@intel.com>