DAOS-16282 control: Fix some test assumptions about JSON
The commit landed for DAOS-16127 resulted in some differences in
JSON output for pool query. Several tests were written in such a
way that the code expected the (en|dis)abled_ranks keys to always
be set, even if those arrays were NULL. This isn't very idiomatic
and is awkward to work with. The test code has been updated to
instead use the get() operator which will return None if the
response dict does not have the requested key.

Also fixes a problem reported in DAOS-16283, where the JSON output
of `dmg pool query` differed from the JSON output of `daos pool query`
because it didn't include the usage array.

Features: pool control
Required-githooks: true

Change-Id: Ia887cee0f3c364e0a9518e179df7494136507b72
Signed-off-by: Michael MacDonald <mjmac@google.com>