DAOS-14679 pool: Backport fixes (#14828)
* DAOS-14679 pool: Report on stopping sp_stopping (#14374)

When destroying a pool on an engine (as part of a pool destroy command),
we may block in

  ds_pool_stop
    pool_put_sync // putting the last reference
      pool_free_ref // called after the hash rec deletion
        pool_child_delete_one
          pool_child_stop

waiting for some ds_pool_child references. If the user retries the pool
destroy command, the new ds_pool_stop call on this engine can't find the
ds_pool object that is in the sp_stopping state, and the new pool
destroy command usually succeeds, even though the storage capacity
allocated to the pool hasn't been released yet.

This patch makes the following changes:

  - Move the pool_child_delete_one collective call from pool_free_ref to
    before the ds_pool_put call in ds_pool_stop. This makes sure that
    the ds_pool object remains in the LRU cache in the sp_stopping state
    while we wait for ds_pool_child references (or some other upper
    layer resource).

  - Remove the pool_put_sync trick, which was introduced because of the
    pool_child_delete_one collective call in pool_free_ref.

  - Return an error from ds_pool_stop when the ds_pool object is in the
    sp_stopping state, so that the new pool destroy command in the
    aforementioned scenario will return an explicit error.

  - Register a reply aggregation callback for MGMT_TGT_DESTROY so that
    an error can reach the control plane.

Signed-off-by: Li Wei <wei.g.li@intel.com>
Required-githooks: true

* DAOS-15998 pool: Let all pool_svc rest on ds_pool (#14681)

Currently, only leader pool_svc objects depend on ds_pool objects. This
unintentionally allows the creation of a PS replica even when the local
ds_pool object is being stopped, leading to the hang in the Jira ticket.
This patch lets every follower pool_svc object depend on the ds_pool
object too, so that if the latter is stopping or doesn't exist, the
pool_svc object won't be created successfully.

  - Initialize pool_svc.ps_pool when allocating pool_svc, rather than
    when stepping up.
      - Move ps_pool up in pool_svc, as it's no longer a leader-only
	field.
      - Modify init_svc_pool accordingly and rename it to
	update_svc_pool.

  - Move ds_pool_svc_stop into ds_pool_stop, because we want to set
    ds_pool.sp_stopping before stopping the PS (if any).
      - Move the PS start code into ds_pool_start for symmetry.
      - Clean up the error handling in and for ds_pool_svc_stop.

  - Change ds_pool_stop_all to stop all pools concurrently.
      - Remove ds_rsvc_stop_all, as it's no longer used anywhere.

Signed-off-by: Li Wei <wei.g.li@intel.com>
Required-githooks: true

* DAOS-16254 pool: Fix rfcheck_ult infinite looping (#14789)

wenbaoxu@163.com found that pool_svc_rfcheck_ult may enter an infinite
loop due to -DER_NOTLEADERs when the PS leader is stepping down:

    src/rdb/rdb_raft.c:3146 rdb_raft_wait_applied() 564e1a52[1]: waiting
      for entry 0 to be applied
    src/container/srv_container.c:5050 ds_cont_rdb_iterate() 564e1a52
      container iter: rc -2008
    src/pool/srv_pool.c:6405 pool_svc_rfcheck_ult() 564e1a52 check rf
      with -2008 and retry
    [...]

This patch adds a check for whether the "sched" is canceled in
pool_svc_rfcheck_ult before sleeping and retrying, and fixes a nonsense
dss_sleep(0) call.


Signed-off-by: Li Wei <wei.g.li@intel.com>