DAOS-16010 control: add support to override rync binary and add args (#14772)
Add a new config (support_config->file_transfer_exec) to override the
default of using rsync to copy files from servers as part of a "collect-log"
command.

  * Added a new option "-T --transfer-args" to pass additional args to the
    executable defined in support_config->file_transfer_exec

This allows an example, to execute an external tool like gsutil to copy
results to a GCS bucket.

Also a fix for allowing the collect command to work in insecure mode is included here.

Signed-off-by: Chris Davis <chrd@google.com>