DAOS-16381 test: Run IOR with HDF5-VOL with multiple object classes (#14964)
Currently, the logic in util/file_count_test_base.py runs IOR with HDF5-VOL
with only one object class. The second (and beyond) object class will not
run with HDF5-VOL (it runs as a normal IOR). Update line 120 to:

if api == 'HDF5-VOL':

so that if multiple object classes are defined in the test yaml, all of them
will run with HDF5-VOL.

Skip-unit-tests: true
Skip-fault-injection-test: true
Skip-func-hw-test-medium-md-on-ssd: false
Test-tag: test_io_sys_admin test_largefilecount test_smallfilecount
Signed-off-by: Makito Kano <makito.kano@intel.com>