DAOS-16478 test: pool/autotest.py - Increase timeout
The two tests in autotest.py with MD-on-SSD timed out in build
248. This was potentially caused by the slow TCP provider.
Build 248 was the only timed out instance in the past 20 builds
(235 to 255), so it looks like an intermittent issue. Increase
the timeout by 50% and run it 10 times and check that it doesn't
time out.

Skip-unit-tests: true
Skip-fault-injection-test: true
Skip-func-hw-test-medium-md-on-ssd: false
Test-tag: test_pool_autotest
Test-repeat: 10

Signed-off-by: Makito Kano <makito.kano@intel.com>