DAOS-16570 control: Remove non-server dependency on libpci (#15131)
Commit 21450930f4 added a new pciutils package that provides
bindings for the library so that the server can interrogate
PCI device status. The work also included a change to the
hwprov package that forces all users of hwprov to pull in
libpci, even though only the server needs it. This commit
changes the server to pull in the pciutils package directly
and removes the dependency from other users of hwprov.

Signed-off-by: Michael MacDonald <mjmac@google.com>