SRE-2505 ci: Trivy scans tuning
- Use GHA cache to avoid Trivy scan failure
Trivy CVEs database downloads fails often.
The most promissing solution is to use cache and download the
database once a day.
CVEs database is cached during daily build (`schedule`).
Cache is not used if `master` branch cache is not available.
https://github.com/aquasecurity/trivy-action?tab=readme-ov-file#updating-caches-in-the-default-branch

- Avoid Trivy scanners re-initialization
https://github.com/aquasecurity/trivy-action?tab=readme-ov-file#skipping-setup-when-calling-trivy-action-multiple-times
The latest available version of `aquasecurity/trivy-action`
is used to be able to use `skip-setup-trivy` parameter.

Doc-only: true

Required-githooks: true

Signed-off-by: Tomasz Gromadzki <tomasz.gromadzki@intel.com>