DAOS-16469 dtx: misc patch for kinds of DTX enhancement - b26
1. shrink DTX table blob size

Use 4KB blob for committed DTX table and 16KB for active DTX table.
It is more efficient for lower allocator and reduce the possibility
of space allocation failure when space pressure.

Simplify vos_dtx_commit logic and code cleanup.

2. optimize DTX CoS cache

If there are a lot of committable DTX entries in DTX CoS cache,
then it may be inefficient to locate the DTX entry in CoS cache
with given oid + dkey_hash, that may happen under the case of
that DTX batched commit is blocked (such as because of network
trouble) as to trigger DTX refresh (for DTX cleanup) on other
related engines. If that happened, it will increase the system
load on such engine and slow down DTX commit further more. The
patch reduces unnecessary search operation inside CoS cache.

Other changes:

2.1. Metrics (io/dtx/async_cmt_lat/tgt_id) for DTX asynchronously
     commit latency (with unit ms).

2.2. Fix a bug in sched_ult2xs() with multiple numa sockets for
     DSS_XS_OFFLOAD case.

2.3. Delay commit (or abort) collective DTX on the leader target
     to handle resent race.

2.4. Avoid blocking dtx_req_wait() if chore failed to send out
     some DTX RPC.

3. properly handle DTX partial commit

When a DTX leader globally commit the DTX, it is possible that some
DTX participant(s) cannot commit such DTX entry because of kinds of
issues, such as network or space trouble. Under such case, the DTX
leader needs to keep the active DTX entry persistently for further
commit/resync. But it does not means related modification attched
to such DTX entry on the leader target cannot be committed, instead,
we can commit related modification with only keeping the DTX header.
That is enough for the DTX leader to do further DTX commit/resync
to handle related former failed DTX participant(s).

The benefit is that VOS aggregation on the leader target will not
be affected by remote DTX commit failure.

The patch also reduces DTX related warning to avoid log flood.

Allow-unstable-test: true

Signed-off-by: Fan Yong <fan.yong@intel.com>