DAOS-15698 test: Update the reset faulty device logic for nvme_utils.py (#15430)
MD-on-SSD cluster: If we set storage faulty, the corresponding engine dies (becomes excluded). Tell the server manager the rank is excluded and don't reset faulty device.
Non-MD-on-SSD cluster: Reset by calling dmg storage led identify --reset. We expect the command to work and all engines to be joined.

Test-tag: test_nvme_fault test_nvme_fault_reintegration test_vmd_led_faulty test_disk_failure_recover
Skip-func-hw-test-medium-vmd: false
Skip-func-hw-test-large-md-on-ssd: false

Signed-off-by: Makito Kano <makito.kano@intel.com>