DAOS-16823 container: Add metrics_enabled property
Add a boolean container property that can be used to
toggle client-side metrics collection and reporting
behavior on a per-container basis. By default, containers
will not participate in metrics collection, and must
be opted-in via the property. The initial use case for
this property is to enable DFS-level metrics in
POSIX containers.

Features: container
Required-githooks: true
Signed-off-by: Michael MacDonald <mjmac@google.com>