DAOS-16807 client: intercept io_queue_init in libpil4dfs (#15497)
Intercepting io_queue_init() is needed on Ubuntu. There is compatibility issue for pil4dfs interception library when used with fio libaio engine. In some cases, fio initialize the aio context through io_queue_init function when loading the libaio engine. Through the pil4dfs has intercepted the io_setup function, but it seems that the io_setup which called by io_queue_init is not intercepted some times, which causing invalid aio context for I/O processing. So add an interception for io_queue_init to make it work for this case.

Signed-off-by: Jun Zeng <jun1.zeng@intel.com>
Signed-off-by: Lei Huang <lei.huang@intel.com>