SRE-2171 ci: Big refactor of GHA workflows (#15556) (#15588)
Add the use of reusable workflows and actions to reduce the amount of
duplicated code in this repository as well as dependency repositories.

Run Bullseye workflow on schedule (#15574)
Saturdays at midnight, UTC.

Accept and propagate a run-gha variable (#15576)
For the case where daos is being used as a downstream test.

Test inputs context before trying to use it.

Fixes: SRE-2570 DAOS-16262

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>