DAOS-16876 vos: set cont parameter when deregister modification from DTX - b26
As long as the container is not destroyed, then anytime want to deregister
a modification from related active DTX entry (that is usually triggered for
vos discard or aggregation), the caller needs to offer container handle to
vos_dtx_deregister_record() for locating the DTX entry in active DTX table.
Otherwise, if the caller offers empty container handle, then it will cause
dangling reference in related DTX entry as to data corruption in subsequent
DTX commit or abort.

On the other hand, if the container will be destroyed, then all related DTX
entries for such container will be useless any more. We need to destroy DTX
table firstly to avoid generating dangling DTX references during destroying
the container.

Allow-unstable-test: true

Signed-off-by: Fan Yong <fan.yong@intel.com>